backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "FastAPI backend for the AgentHub MVP registry and execution APIs"
readme = "README.md"
authors = [
    { name = "Oliver Jay", email = "oliver.tj@oliver.tj" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "fastmcp>=3.2.0",
    "jsonschema>=4.26.0",
    "pydantic-ai-slim[openai]>=0.0.55",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "pydantic-ai>=1.78.0",
    "web3>=7.13.0",
]

[project.scripts]
backend = "backend.cli:main"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.basedpyright]
include = ["src", "tests"]
exclude = ["alembic", "typings", ".venv", ".pytest_cache", "**/__pycache__"]
extraPaths = ["src"]
stubPath = "typings"
pythonVersion = "3.14"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"




[tool.mypy]
files = ["src", "tests"]
mypy_path = "src:typings"
python_version = "3.14"
warn_unused_ignores = true
warn_redundant_casts = true
strict_equality = true
exclude = ["alembic"]

[build-system]
requires = ["uv_build>=0.11.4,<0.12.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "basedpyright>=1.39.0",
    "httpx>=0.28.1",
    "mypy>=1.20.0",
    "pytest>=9.0.3",
]
